feat: doctest everywhere
