+++ pyproject.toml
@@ -8,11 +8,14 @@ readme = "README.md"
 homepage = "https://github.com/clintval/typeline"
 repository = "https://github.com/clintval/typeline"
 keywords = ["dataclass", "msgspec", "IO", "delimited", "CSV", "TSV"]
+include = ["CONTRIBUTING.md", "LICENSE"]
 classifiers = [
     "Development Status :: 3 - Alpha",
     "Environment :: Console",
     "Intended Audience :: Developers",
+    "Intended Audience :: Science/Research",
     "License :: OSI Approved :: MIT License",
+    "Natural Language :: English",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.10",
@@ -23,24 +26,22 @@ classifiers = [
     "Topic :: Software Development :: Libraries :: Python Modules",
     "Typing :: Typed",
 ]
-include = ["CONTRIBUTING.md", "LICENSE"]
 
 [tool.poetry.dependencies]
-python = "^3.10.0"
-msgspec = "^0.18"
-typing-extensions = "^4.12"
+python = "^3.10.0,<4.0.0"
+msgspec = "^0.18,<1.0"
+typing-extensions = "^4.12,<5.0"
 
 [tool.poetry.dev-dependencies]
 basedpyright = "^1.21"
 mypy = "^1.13"
 pytest = "^8.3"
 pytest-cov = "^5.0"
-# TODO: get doctests to work and also run on the README?
-pytest-doctestplus = "^1.2.1"
-ruff = "0.7.1"
+pytest-doctestplus = "^1.2"
+ruff = "^0.7"
 
 [build-system]
-requires = ["poetry-core>=1.8", "setuptools>=70.1.1"]
+requires = ["poetry-core>=1.8"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poe]
@@ -56,13 +57,17 @@ fix-all.sequence    = [
     "fix-lint"
 ]
 
-check-lock     = "poetry check --lock"
-check-format   = "ruff format --check --diff"
-check-lint     = "ruff check"
-check-tests    = "pytest"
+check-lock   = "poetry check --lock"
+check-format = "ruff format --check --diff"
+check-lint   = "ruff check"
+check-tests  = "pytest --doctest-glob='*.md'"
+
 _check-mypy    = "mypy"
 _check-pyright = "basedpyright"
-check-typing.sequence = ["_check-mypy", "_check-pyright"]
+check-typing.sequence = [
+    "_check-mypy",
+    "_check-pyright"
+]
 
 check-all.ignore_fail = true
 check-all.sequence    = [
@@ -77,6 +82,25 @@ fix-and-check-all.ignore_fail = true
 fix-and-check-all.sequence    = [
     "fix-all",
     "check-all"
+]
+
+[tool.pytest.ini_options]
+minversion = "7.4"
+addopts    = [
+    "--color=yes",
+    "--cov-fail-under=80",
+    "--cov-report=term-missing",
+    "--cov",
+    "--cov=typeline",
+    "--doctest-modules",
+    "--doctest-plus",
+    "--import-mode=importlib",
+]
+doctest_plus = "enabled"
+doctest_optionflags = [
+    "ELLIPSIS",
+    "IGNORE_EXCEPTION_DETAIL",
+    "NORMALIZE_WHITESPACE",
 ]
 
 [tool.mypy]
@@ -97,18 +121,14 @@ warn_unused_configs         = true
 warn_unused_ignores         = true
 enable_error_code           = "ignore-without-code"
 
+[[tool.mypy.overrides]]
+module = ".conftest"
+ignore_missing_imports = true
+
 [tool.pyright]
 include = ["typeline/"]
 reportAny = false
 pythonVersion = "3.11"
-
-[tool.pytest.ini_options]
-minversion = "7.4"
-addopts    = [
-    "--color=yes",
-    "--import-mode=importlib",
-    "--cov"
-]
 
 [tool.ruff]
 include        = ["typeline/**", "tests/**"]
@@ -141,6 +161,9 @@ ignore    = [
 ]
 unfixable = ["B"]
 
+[tool.ruff.lint.mccabe]
+max-complexity = 10
+
 [tool.ruff.lint.isort]
 force-single-line = true
 
@@ -163,6 +186,7 @@ body = """
 conventional_commits = true
 commit_parsers = [
     { message = "^.+!:*", group = "Breaking"},
+    { message = "^security*", group = "Security"},
     { message = "^feat*", group = "Features"},
     { message = "^fix*", group = "Bug Fixes"},
     { message = "^docs*", group = "Documentation"},
